chore: 抽离dom操作
